# Performance optimization of LayerNorm on CPU