@pengzhao-intel 2019-06-21
Thanks for the improvements. Is there a test case to cover this OP?