@TaoLv 2019-06-21
@pengxin99 Thank you for the contribution. Could you change the title and description? I think this PR and the LayerNorm operator is not related to MKL-DNN. Better to call it "Performance optimization of LayerNorm on CPU".

@pengzhao-intel Test should be covered by https://github.com/apache/incubator-mxnet/blob/master/tests/python/unittest/test_operator.py#L3594 and https://github.com/apache/incubator-mxnet/blob/master/tests/python/unittest/test_gluon.py#L741.